Initial project setup
This commit introduces the basic structure of the MauiBlazorNotification project. It includes the establishment of the necessary services, the configuration of the MainActivity, and the implementation of Firebase token retrievals and notifications in a Blazor-based Maui application. Detailed documentation is provided in the README.md file, explaining the structure of the project, dependencies, and steps for local setup.

Please note that this is just the initial commit, and further improvements and feature additions may follow. We welcome constructive feedback and contributions from the community to help enhance this project.